feat(dotcom): sidebar polish and dialog dismiss fixes
Extract TlaSidebarWorkspaceActions into its own component and render it from
TlaSidebar. Let portalled controls opt out of dialog dismissal via a
data-tlui-dialog-prevent-dismiss attribute instead of disabling background close
for the whole dialog, so interacting with a TlaMenuSelect popup no longer closes
the workspace settings dialog. Add right-click support to the user settings menu,
and blur (rather than refocus) the sidebar search input when clearing it.
