+++ apps/dotcom/client/src/tla/components/tla-menu/tla-menu.tsx
@@ -188,6 +188,7 @@ export function TlaMenuSelect<T extends string>({
 				<TlaSelectPortal usePortal={usePortal} container={container}>
 					<_Select.Content
 						className={styles.menuSelectContent}
+						data-tlui-dialog-prevent-dismiss
 						position={usePortal ? 'popper' : undefined}
 						side={usePortal ? 'bottom' : undefined}
 						align={usePortal ? 'end' : undefined}
